requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-decouple>=3.8
psycopg2-binary>=2.9.6  # PostgreSQL adapter (optional, remove if using SQLite)
djangorestframework-simplejwt>=5.2.2  # JWT tokens (optional)
